remove previous google sheets code
